Add bulk actions as default on active offers page
