+++ src/pages/OffersPage/components/ActiveTabContent/components/LoansTable/columns.tsx
@@ -3,8 +3,8 @@ import { ColumnType } from '@banx/components/Table'
 import { HeaderCell, NftInfoCell, createSolValueJSX } from '@banx/components/TableComponents'
 
 import { Loan } from '@banx/api/core'
-import { calculateLentValue } from '@banx/pages/OffersPage'
-import { formatDecimal, isLoanTerminating } from '@banx/utils'
+import { calculateLentValue, isLoanAbleToTerminate } from '@banx/pages/OffersPage'
+import { formatDecimal } from '@banx/utils'
 
 import { APRCell, ActionsCell, InterestCell, StatusCell } from './TableCells'
 
@@ -15,7 +15,6 @@ interface GetTableColumnsProps {
   toggleLoanInSelection: (loan: Loan) => void
   onSelectAll: () => void
 
-  isUnderwaterFilterActive: boolean
   hasSelectedLoans: boolean
   isCardView: boolean
 }
@@ -24,7 +23,6 @@ export const getTableColumns = ({
   findLoanInSelection,
   toggleLoanInSelection,
   onSelectAll,
-  isUnderwaterFilterActive,
   hasSelectedLoans,
   isCardView,
 }: GetTableColumnsProps) => {
@@ -33,25 +31,14 @@ export const getTableColumns = ({
       key: 'collateral',
       title: (
         <div className={styles.headerTitleRow}>
-          {isUnderwaterFilterActive && (
-            <Checkbox
-              className={styles.checkbox}
-              onChange={onSelectAll}
-              checked={hasSelectedLoans}
-            />
-          )}
+          <Checkbox className={styles.checkbox} onChange={onSelectAll} checked={hasSelectedLoans} />
           <HeaderCell label="Collateral" align="left" />
         </div>
       ),
       render: (loan) => {
         const { partnerPoints = 0, playerPoints = 0, name, imageUrl } = loan.nft.meta
 
-        const canSelect = isUnderwaterFilterActive && !isLoanTerminating(loan)
-
-        const onCheckboxClick = isUnderwaterFilterActive
-          ? () => toggleLoanInSelection(loan)
-          : undefined
-
+        const canSelect = isLoanAbleToTerminate(loan)
         const selected = canSelect ? !!findLoanInSelection(loan.publicKey) : undefined
 
         return (
@@ -59,7 +46,7 @@ export const getTableColumns = ({
             nftName={name}
             nftImage={imageUrl}
             selected={selected}
-            onCheckboxClick={onCheckboxClick}
+            onCheckboxClick={() => toggleLoanInSelection(loan)}
             banxPoints={{ partnerPoints, playerPoints }}
             checkboxClassName={!canSelect ? styles.nftCellCheckbox : ''}
           />
